# server/controllers/Choonti_db/generated/ShopControllerGenerated.js
/* 
* Generated by
* 
*      _____ _          __  __      _     _
*     / ____| |        / _|/ _|    | |   | |
*    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
*     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
*     ____) |   < (_| | | | || (_) | | (_| |  __/ |
*    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
*
* The code generator that works in many programming languages
*
*			https://www.skaffolder.com
*
*
* You can generate the code from the command-line
*       https://npmjs.com/package/skaffolder-cli
*
*       npm install -g skaffodler-cli
*
*   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
*
* To remove this comment please upgrade your plan here: 
*      https://app.skaffolder.com/#!/upgrade
*
* Or get up to 70% discount sharing your unique link:
*       https://app.skaffolder.com/#!/register?friend=5d6cf8a94a4f7c744b6c964c
*
* You will get 10% discount for each one of your friends
* 
*/
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE ShopControllerGenerated.js PLEASE EDIT ../ShopController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import ShopModel from "../../../models/Choonti_db/ShopModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import ShopController from "../ShopController";
const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/shop`;
    router.post(baseUrl + "", authorize([]), ShopController.create);
    router.delete(baseUrl + "/:id", authorize([]), ShopController.delete);
    router.get(baseUrl + "/findByarea/:key", authorize([]), ShopController.findByarea);
    router.get(baseUrl + "/:id", authorize([]), ShopController.get);
    router.get(baseUrl + "", authorize([]), ShopController.list);
    router.post(baseUrl + "/:id", authorize([]), ShopController.update);
  },


  // CRUD METHODS


  /**
  * ShopModel.create
  *   @description CRUD ACTION create
  *   @param Shop obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await ShopModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ShopModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id Shop
  *
  */
  delete: async (req, res) => {
    try {
      const result = await ShopModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ShopModel.findByarea
  *   @description CRUD ACTION findByarea
  *   @param Objectid key Id of the resource area to search
  *
  */
  findByarea: async (req, res) => {
    try {
      const result = await ShopModel.findByarea(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ShopModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id Shop
  *   @returns Shop
  *
  */
  get: async (req, res) => {
    try {
      const result = await ShopModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ShopModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF Shop
  *
  */
  list: async (req, res) => {
    try {
      const result = await ShopModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * ShopModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id Shop
  *   @returns Shop
  *
  */
  update: async (req, res) => {
    try {
      const result = await ShopModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;